+++ R/significance.R
@@ -66,6 +66,7 @@ calculate_significance <- function(hojas_list, sufijo, type, main_var_prop, des_
   est_prefix <- switch(type,
     prop = "prop",
     mean = "media",
+    total = "total",
     quantile = "cuantil"
   )
   se_prefix  <- "se"